nlp-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-cloud-aiplatform==1.40.0
nltk==3.8.1
python-multipart==0.0.6
tenacity==8.2.3
python-json-logger==2.0.7
prometheus-client==0.20.0
python-dotenv==1.0.1